refactor guardian code into more organized module
