+++ ethereum/network.py
@@ -1,7 +1,28 @@
 import random
-import sys
 import time
-from utils import DEBUG
+
+import rlp
+import gevent
+
+from ethereum.guardian2.network import (
+    NetworkMessage,
+    GuardianService,
+    GuardianProtocol,
+    GuardianApp,
+)
+
+
+from devp2p import app_helper
+from devp2p import peermanager
+from devp2p.discovery import NodeDiscovery
+from devp2p.crypto import (
+    privtopub as privtopub_raw,
+)
+from devp2p.utils import (
+    host_port_pubkey_to_uri,
+    update_config_with_defaults,
+)
+
 
 # A network simulator
 
@@ -227,172 +248,10 @@ class SimPyNetworkSimulator(NetworkSimulatorBase):
 #NetworkSimulator = SimPyNetworkSimulator
 
 
-import sys
-import random
-import copy
-import six
-import rlp
-import gevent
-
-from devp2p.crypto import (
-    sha3,
-    privtopub as privtopub_raw,
-)
-from devp2p.app import BaseApp
-from devp2p.protocol import BaseProtocol
-from devp2p.service import WiredService
-from devp2p.utils import (
-    colors,
-    COLOR_END,
-    host_port_pubkey_to_uri,
-    update_config_with_defaults,
-)
-from devp2p import app_helper
-
-import ethereum.slogging as slogging
-from ethereum.guardian import NetworkMessage
-
-
-#slogging.configure(config_string=':debug,p2p.discovery:info')
-slogging.configure(config_string=':info,p2p.discovery:info')
-logger = slogging.get_logger('guardian')
-
-_nonce = 0
-
-
-#
-# Protocol
-#
-class GuardianProtocol(BaseProtocol):
-    protocol_id = 1
-    network_id = 0
-    max_cmd_id = 1  # Actually max id is 0, but 0 is the special value.
-    name = 'guardian'
-    version = 1
-
-    def __init__(self, peer, service):
-        # required by P2PProtocol
-        self.config = peer.config
-        BaseProtocol.__init__(self, peer, service)
-
-    class network_message(BaseProtocol.command):
-        """
-        A guardian network message
-        """
-        cmd_id = 0
-
-        structure = [
-            ('network_message', NetworkMessage)
-        ]
-
-#
-#  Service
-#
-class GuardianService(WiredService):
-
-    # required by BaseService
-    name = 'guardianservice'
-    default_config = {
-        'guardianservice': {
-            'num_participants': 1,
-        }
-    }
-
-    # required by WiredService
-    wire_protocol = GuardianProtocol
-
-    def __init__(self, app):
-        self.config = app.config
-        self.address = privtopub_raw(self.config['node']['privkey_hex'].decode('hex'))
-        super(GuardianService, self).__init__(app)
-
-    def log(self, text, **kargs):
-        node_num = self.config['node_num']
-        msg = ' '.join([
-            colors[node_num % len(colors)],
-            "NODE%d" % node_num,
-            text,
-            (' %r' % kargs if kargs else ''),
-            COLOR_END])
-        logger.debug(msg)
-
-    def broadcast(self, obj, origin=None):
-        fmap = {NetworkMessage: 'network_message'}
-        self.log('broadcasting', obj=obj)
-        bcast = self.app.services.peermanager.broadcast
-        bcast(GuardianProtocol, fmap[type(obj)], args=(obj,),
-              exclude_peers=[origin.peer] if origin else [])
-
-    def on_wire_protocol_stop(self, proto):
-        assert isinstance(proto, self.wire_protocol)
-        self.log('----------------------------------')
-        self.log('on_wire_protocol_stop', proto=proto)
-
-    def on_wire_protocol_start(self, proto):
-        self.log('----------------------------------')
-        self.log('on_wire_protocol_start', proto=proto, peers=self.app.services.peermanager.peers)
-        assert isinstance(proto, self.wire_protocol)
-        # register callbacks
-        proto.receive_network_message_callbacks.append(self.on_receive_network_message)
-
-    def on_receive_network_message(self, proto, network_message):
-        assert isinstance(network_message, NetworkMessage)
-        assert isinstance(proto, self.wire_protocol)
-        self.log('----------------------------------')
-        self.log('on_receive network_message', network_message=network_message, proto=proto)
-        agent = self.app.services.guardianservice.config['guardianservice']['agent']
-        lookup_fn = self.app.services.guardianservice.config['guardianservice']['lookup_fn']
-        other_agent_id = lookup_fn(proto.peer.remote_pubkey)
-        agent.on_receive(rlp.encode(network_message), other_agent_id)
-
-    def send_network_message(self, typ, args):
-        # without this we hit a race condition between our nodes.
-        gevent.sleep(random.random())
-        network_message = NetworkMessage(typ=typ, args=args)
-        self.log('----------------------------------')
-        self.log('sending network_message', network_message=network_message)
-        self.broadcast(network_message)
-
-
-#
-#  App
-#
-guardian_client_name = "guardianapp"
-guardian_version = '0.1'
-guardian_client_version = ''.join((
-    guardian_version,
-    sys.platform,
-    'py{:d}.{:d}.{:d}'.format(*sys.version_info[:3]),
-))
-guardian_client_version_string = '{:s}/v{:s}'.format(
-    guardian_client_name,
-    guardian_client_version,
-)
-
-guardian_default_config = copy.deepcopy(BaseApp.default_config)
-guardian_default_config.update({
-    'client_version_string': guardian_client_version_string,
-    'post_app_start_callback': None,
-})
-
-
-class GuardianApp(BaseApp):
-    client_name = guardian_client_name
-    version = guardian_version
-    client_version = guardian_client_version
-    client_version_string = guardian_client_version_string
-
-    default_config = guardian_default_config
-
-
-from devp2p import app_helper
-from devp2p import peermanager
-from devp2p.service import BaseService
-from devp2p.discovery import NodeDiscovery
-
-
 class DevP2PNetwork(NetworkSimulatorBase):
-
+    """
+    *Mostly* drop in networking object for use with `ethereum/test.py` script.
+    """
     start_time = time.time()
 
     def __init__(self, agents=None, seed=0, min_peers=2, max_peers=5, random_port=False):
@@ -491,13 +350,6 @@ class DevP2PNetwork(NetworkSimulatorBase):
             exclude_peers=[],
         )
 
-        #if random.random() < self.broadcast_success_rate:
-        #    for p in self.peers[sender.id]:
-        #        recv_time = self.time + self.latency_distribution_sample()
-        #        if recv_time not in self.objqueue:
-        #            self.objqueue[recv_time] = []
-        #        self.objqueue[recv_time].append((sender.id, p, obj))
-
     def send_to_one(self, sender, obj):
         assert isinstance(obj, (str, bytes))
 
@@ -519,7 +371,6 @@ class DevP2PNetwork(NetworkSimulatorBase):
                 break
 
         if to_peer is None:
-            import ipdb; ipdb.set_trace()
             raise ValueError("Not connected to the provided agent")
 
         proto = to_peer.protocols[GuardianProtocol]
